Added WIP files
bigDataSet is the modified version of the codes from: https://www.tensorflow.org/tutorials/text/word2vec. It behaves in the same manner and generates the same output as the tutorial's codes.

main.py is my codes that used a much smaller dataset. 

Main Bug to work on is the lack of samples when my codes are trying to generate skipgrams